requirements.txt
flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
yfinance==0.2.31
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
fredapi==0.5.1
requests==2.31.0
httpx>=0.25.0
apscheduler==3.10.4
odfpy>=1.4.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
